requirements.txt
# Webframework
fastapi[standard-no-fastapi-cloud-cli]
uvicorn[standard]

# Database & ORM
sqlalchemy>=2.0
psycopg2-binary
alembic

# Data validation
pydantic>=2.0
pydantic-settings

# Environment & Config
python-dotenv

# Security
bcrypt
python-jose[cryptography]
passlib

# Async & Task Queuing
celery>=5.3.0
redis>=5.0.0
kombu>=5.3.0
watchdog>=3.0.0  # Required for Celery autoreload

# LLM & AI
langchain>=0.1.0
langchain-openai>=0.0.1
openai>=1.0.0
langgraph>=0.0.1

# Cloud & Storage
boto3>=1.26.0  # AWS S3 and S3-compatible storage (IONOS S3)

# Testing
pytest>=7.0
pytest-asyncio
httpx

# Logging & Monitoring
structlog==23.2.0

# Utilities
python-dateutil